+++ contracts/cw-ics20-latest/src/testing/mod.rs
@@ -1,0 +1,4 @@
+mod ibc_hooks_test;
+mod ibc_tests;
+mod integration_tests;
+mod test_helpers;
