+++ contracts/cw-ics20-latest/src/testing/test_helpers.rs
@@ -1,5 +1,3 @@
-#![cfg(test)]
-
 use crate::contract::instantiate;
 use crate::ibc::{ibc_channel_connect, ibc_channel_open, ICS20_ORDERING, ICS20_VERSION};
 use crate::state::ChannelInfo;
@@ -18,7 +16,7 @@ pub const DEFAULT_TIMEOUT: u64 = 3600; // 1 hour,
 pub const CONTRACT_PORT: &str = "wasm.cosmos2contract"; // wasm.MOCK_CONTRACT_ADDR
 pub const REMOTE_PORT: &str = "transfer";
 pub const CONNECTION_ID: &str = "connection-2";
-pub const WASM_BYTES: &[u8] = include_bytes!("../artifacts/cw-ics20-latest.wasm");
+pub const WASM_BYTES: &[u8] = include_bytes!("../../artifacts/cw-ics20-latest.wasm");
 
 pub fn mock_channel(channel_id: &str) -> IbcChannel {
     IbcChannel::new(
