+++ packages/cw20-ics20-msg/Cargo.toml
@@ -15,10 +15,8 @@ cosmwasm-schema = "=1.2"
 cosmwasm-std = { version = "=1.2", default-features = false }
 cw-storage-plus = "1.0.1"
 cw20 = "1.0.1"
-schemars = "0.8.1"
 bech32 = "0.8.1"
-oraiswap = { git = "https://github.com/oraichain/oraiswap.git", rev = "03f4955b" }
-serde = { version = "1.0.103", default-features = false, features = ["derive"] }
+oraiswap = { workspace = true }
 anybuf = "0.3.0"
 
 [dev-dependencies]
