+++ packages/cw20-ics20-msg/src/ibc_hooks.rs
@@ -77,7 +77,7 @@ impl IbcHooksUniversalSwap {
 }
 
 #[cfg(test)]
-mod test {
+mod tests {
 
     use anybuf::Anybuf;
     use cosmwasm_std::{Api, Binary, StdError};
